Updating public repo as per the latest develop
