+++ samples/configs/hpc_catalog_values.json
@@ -20,9 +20,9 @@
     "enable_cos_integration"    : "false",
     "cos_instance_name"         : "__NULL__",
     "enable_fip"                : "true",
-    "management_image_name"     : "hpcaas-lsf10-rhel88-v9",
-    "compute_image_name"        : "hpcaas-lsf10-rhel88-compute-v5",
-    "login_image_name"          : "hpcaas-lsf10-rhel88-compute-v5",
+    "management_image_name"     : "hpcaas-lsf10-rhel88-v10",
+    "compute_image_name"        : "hpcaas-lsf10-rhel88-compute-v6",
+    "login_image_name"          : "hpcaas-lsf10-rhel88-compute-v6",
     "login_node_instance_type"  : "bx2-2x8",
     "management_node_instance_type" : "bx2-16x64",
     "management_node_count"     : "3",
