+++ tests/lsf/constants.go
@@ -11,13 +11,13 @@ const (
 	LSF_CUSTOM_RESOURCE_GROUP_VALUE_AS_NULL = "null"
 	LOGIN_NODE_EXECUTION_PATH               = "source /opt/ibm/lsf/conf/profile.lsf;"
 	COMPUTE_NODE_EXECUTION_PATH             = "source /opt/ibm/lsf_worker/conf/profile.lsf;"
-	JOB_COMMAND_LOW_MEM                     = `bsub -J myjob[1-2] -R "select[family=mx2] rusage[mem=10G]" sleep 90`
-	JOB_COMMAND_MED_MEM                     = `bsub -J myjob[1-2] -R "select[family=mx2] rusage[mem=30G]" sleep 90`
-	JOB_COMMAND_HIGH_MEM                    = `bsub -J myjob[1-2] -R "select[family=mx2] rusage[mem=90G]" sleep 90`
-	JOB_COMMAND_LOW_MEM_SOUTH               = `bsub -J myjob[1-2] -R "select[family=mx3d] rusage[mem=10G]" sleep 90`
-	JOB_COMMAND_MED_MEM_SOUTH               = `bsub -J myjob[1-2] -R "select[family=mx3d] rusage[mem=30G]" sleep 90`
-	JOB_COMMAND_HIGH_MEM_SOUTH              = `bsub -J myjob[1-2] -R "select[family=mx3d] rusage[mem=90G]" sleep 90`
-	JOB_COMMAND_LOW_MEM_WITH_MORE_SLEEP     = `bsub -J myjob[1-2] -R "select[family=mx2] rusage[mem=30G]" sleep 90`
+	JOB_COMMAND_LOW_MEM                     = `bsub -J myjob[1-1] -R "select[family=mx2] rusage[mem=10G]" sleep 90`
+	JOB_COMMAND_MED_MEM                     = `bsub -J myjob[1-1] -R "select[family=mx2] rusage[mem=30G]" sleep 90`
+	JOB_COMMAND_HIGH_MEM                    = `bsub -J myjob[1-1] -R "select[family=mx2] rusage[mem=90G]" sleep 90`
+	JOB_COMMAND_LOW_MEM_SOUTH               = `bsub -J myjob[1-1] -R "select[family=mx3d] rusage[mem=10G]" sleep 90`
+	JOB_COMMAND_MED_MEM_SOUTH               = `bsub -J myjob[1-1] -R "select[family=mx3d] rusage[mem=30G]" sleep 90`
+	JOB_COMMAND_HIGH_MEM_SOUTH              = `bsub -J myjob[1-1] -R "select[family=mx3d] rusage[mem=90G]" sleep 90`
+	JOB_COMMAND_LOW_MEM_WITH_MORE_SLEEP     = `bsub -J myjob[1-1] -R "select[family=mx2] rusage[mem=30G]" sleep 90`
 )
 
 var (
